Store encoded ASN.1 for further use
It can be useful to store the encoding for future use, for example to verify a
signature of encoded type without re-encoding (which can fail due to different
encodings).

Note: there may be a more efficient way to do this that stores a view into the
Data rather than copying.